// frontend/src/app/models/user.model.ts
export class user {
    id?: any;
    personnel_fname?: string;
    personnel_lname?: string;
    personnel_caedID?: string;
    personnel_phone?: string;
    personnel_email?: string;
    personnel_username?: string;
    personnel_password?: string;
    personnel_img?: string;
    personnel_role?: string;
  }